Created a file service with method for reading lines of a file. Fixed messages while checking csv validity.
