+++ historian/frontend/src/app/modules/tag/tag-import/tag-csv-import.component.ts
@@ -5,6 +5,7 @@ import { FormGroup, FormBuilder, Validators, AbstractControl } from '@angular/fo
 import { QuestionBase } from '../../../shared/dynamic-form/question-base';
 import { TextboxQuestion } from '../../../shared/dynamic-form/question-textbox';
 import { parse, ParseConfig, ParseResult } from 'papaparse';
+import { FileUtil } from '../../../shared/file/file.service';
 
 @Component({
   selector: 'app-tag-csv-import',
@@ -16,6 +17,9 @@ export class TagCsvImportComponent implements OnInit {
   form: FormGroup;
   private separatorCtrl: AbstractControl;
   private encodingCtrl: AbstractControl;
+  private separatorQuestion: TextboxQuestion;
+  private encodingQuestion: TextboxQuestion;
+
 
   questions: QuestionBase<any>[];
   headers: Set<IHeader>;
@@ -35,27 +39,30 @@ export class TagCsvImportComponent implements OnInit {
   };
 
   constructor(private tagHistorianService: TagHistorianService,
-              private fb: FormBuilder) {
+              private fb: FormBuilder,
+              private fileUtil: FileUtil) {
     this.form = this.fb.group({
       content: [null, Validators.required],
-      separator: [';', Validators.required],
+      separator: [',', Validators.required],
       encoding: ['UTF-8', Validators.required],
     });
     this.separatorCtrl = this.form.get('separator');
     this.encodingCtrl = this.form.get('encoding');
+    this.separatorQuestion = new TextboxQuestion({
+      key: 'separator',
+      label: 'Csv delimiter',
+      order: 1,
+      required: true,
+    });
+    this.encodingQuestion = new TextboxQuestion({
+      key: 'encoding',
+      label: 'Csv encoding',
+      order: 2,
+      required: true,
+    });
     this.questions = [
-      new TextboxQuestion({
-        key: 'separator',
-        label: 'Csv delimiter',
-        order: 1,
-        required: true,
-      }),
-      new TextboxQuestion({
-        key: 'encoding',
-        label: 'Csv encoding',
-        order: 2,
-        required: true,
-      }),
+      this.separatorQuestion,
+      this.encodingQuestion    
     ];
    }
 
@@ -75,7 +82,14 @@ export class TagCsvImportComponent implements OnInit {
 
   private selectFile(file: File): void {
     this.currentFile = file;
+    this.resetMsgs();
+    this.setQuestionEditable();
     this.fileSizeString = this.calculStringSize(file);
+  }
+
+  private setQuestionEditable(): void {
+    this.encodingQuestion.readonly = false;
+    this.separatorQuestion.readonly = false;
   }
 
   private resetMsgs(): void {
@@ -90,7 +104,7 @@ export class TagCsvImportComponent implements OnInit {
       encoding: this.encodingCtrl.value,
       delimiter: this.separatorCtrl.value
     };
-    this.readSomeLines(this.currentFile, 1,
+    this.fileUtil.readSomeLines(this.currentFile, 1,
       line => this.validateCsvHeader(line, parseConfig),
       () => console.log('Completed'),
       this.encodingCtrl.value,
@@ -115,88 +129,31 @@ export class TagCsvImportComponent implements OnInit {
         missingHeaders.push(h.name);
       }
     });
-    if (missingHeaders.length === 0) {
+    if (missingHeaders.length === 0) {//VALID      
       this.displaySuccessMsg = true;
       this.missingHeaders = missingHeaders;
-    } else {
-      this.displayErrMsg = true;
+      this.encodingQuestion.readonly = true;
+      this.separatorQuestion.readonly = true;      
+    } else {//NOT VALID
+      this.displayErrMsg = true;      
       this.missingHeaders = missingHeaders;
     }
     this.headerCurrentFile = headers;
     this.validating = false;
   }
 
-  /**
-   * Read up to and including |maxlines| lines from |file|.
-   *
-   * @param {File} file - The file to be read.
-   * @param {integer} maxlines - The maximum number of lines to read.
-   * @param {function(string)} forEachLine - Called for each line.
-   * @param {function(error)} onComplete - Called when the end of the file
-   *     is reached or when |maxlines| lines have been read.
-   */
-  private readSomeLines(file: File, maxlines: number, forEachLine, onComplete,
-                        encoding?: string, decoderOptions?: TextDecoderOptions) {
-    const CHUNK_SIZE = 50000; // 50kb, arbitrarily chosen.
-    const textDecoder: TextDecoderOptions = {
-      fatal: false,
-      ignoreBOM: false
-    };
-    if (decoderOptions) {
-      Object.assign(textDecoder, decoderOptions);
-    }
-    const decoder = new TextDecoder(encoding || 'utf-8', textDecoder);
-    let offset = 0;
-    let linecount = 0;
-    let results = '';
-    const fr = new FileReader();
-    fr.onload = function() {
-        // Use stream:true in case we cut the file
-        // in the middle of a multi-byte character
-        results += decoder.decode(fr.result, {stream: true});
-        const lines = results.split('\n');
-        results = lines.pop(); // In case the line did not end yet.
-        linecount += lines.length;
+  
 
-        if (linecount > maxlines) {
-            // Read too many lines? Truncate the results.
-            lines.length -= linecount - maxlines;
-            linecount = maxlines;
-        }
-
-        for (let i = 0; i < lines.length; ++i) {
-            forEachLine(lines[i] + '\n');
-        }
-        offset += CHUNK_SIZE;
-        seek();
-    };
-    fr.onerror = function() {
-        onComplete(fr.error);
-    };
-    seek();
-
-    function seek() {
-      if (linecount === maxlines) {
-          // We found enough lines.
-          onComplete(); // Done.
-          return;
-      }
-      if (offset !== 0 && offset >= file.size) {
-          // We did not find all lines, but there are no more lines.
-          forEachLine(results); // This is from lines.pop(), before.
-          onComplete(); // Done
-          return;
-      }
-      const slice = file.slice(offset, offset + CHUNK_SIZE);
-      fr.readAsArrayBuffer(slice);
-    }
-  }
-
-  importCsv() {
+  importCsv(file: File) {
 
     this.progress.percentage = 0;
+    
+    this.tagHistorianService.importTagCsv(file, {
+      separator: this.separatorCtrl.value,
+      charset: this.encodingCtrl.value,
+      bulkSize: 10000
+    })
 
-    // this.currentFile = this.selectedFiles.item(0);
     // this.uploadService.pushFileToStorage(this.currentFileUpload).subscribe(event => {
     //   if (event.type === HttpEventType.UploadProgress) {
     //     this.progress.percentage = Math.round(100 * event.loaded / event.total);
