+++ historian/frontend/src/app/shared/dynamic-form/question-helper.ts
@@ -2,6 +2,11 @@ import { QuestionBase, IQuestionBase } from './question-base';
 
 export class Questions {
 
+  /**
+   * 
+   * @param questions to modify
+   * @param mododifications modifications for each question (property key is used)
+   */
     public static modifyQuestions(
         questions: QuestionBase<any>[],
         mododifications: Map<string, IQuestionBase<any>>
